diabetes_backend/requirements.txt
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
eventlet==0.39.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
PyJWT==2.10.1
pymongo==4.12.1
PyMuPDF==1.25.5
pytesseract==0.3.13
python-dotenv==1.1.0
python-engineio==4.12.0
python-socketio==5.13.0
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.4.0
Werkzeug==3.1.3
wsproto==1.2.0

langchain==0.2.1
chromadb==0.4.24
sentence-transformers==2.6.1
